Move filter to directives folder; Implement filter emitters
