+++ demos/app/app.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 import { BrowserModule } from "@angular/platform-browser";
 import { FormsModule } from "@angular/forms";
-import { FilterModule } from '../../src/list/filter';
+import { FilterModule } from '../../src/directives/filter';
 
 import {
